+++ news/pep517.vendor.rst
@@ -1,0 +1,1 @@
+Upgrade pep517 to 0.10.0
