+++ src/pip/_vendor/pep517/__init__.py
@@ -1,6 +1,6 @@
 """Wrappers to build Python packages using PEP 517 hooks
 """
 
-__version__ = '0.9.1'
+__version__ = '0.10.0'
 
 from .wrappers import *  # noqa: F401, F403
