+++ src/pip/_vendor/resolvelib.pyi
@@ -1,0 +1,1 @@
+from resolvelib import *
