+++ src/pip/_vendor/vendor.txt
@@ -6,7 +6,7 @@ distro==1.5.0
 html5lib==1.1
 msgpack==1.0.2
 packaging==20.9
-pep517==0.9.1
+pep517==0.10.0
 progress==1.5
 pyparsing==2.4.7
 requests==2.25.1
